add style changes for dark theme
